refactor: move limits module to utils/limits
Move limits from tinyagent/limits to tinyagent/utils/limits to organize
utility modules under a common directory.

- Remove tinyagent/limits/ (duplicate of utils/limits)
- Update imports in __init__.py, agents/code.py, execution/local.py
- Verified identical content already exists in utils/limits/

Fixes #13
